perf(hist): float histogram cells, double reductions (decision 50)
HistCell is {float, float}: per-cell sums are bounded by node size and
gradients arrive as float, so cell storage carries only per-cell rounding.
Reductions that cross cells — totals(), fill_prefix, the node finder's
running left sums — accumulate in double and convert once at the store.
The CUDA engine already proved the shape (float per-chunk, double merge).

Paired on M2 (8 threads): base cell fit 15.3 -> 13.2s, 8M x 100 53.3 ->
45.7s (populate 1.23-1.25x); R2 identical to six decimals at both scales;
eval baseline moved 5.4e-6 (re-pinned). Histogram pool, partial slabs, and
prefix scratch all halve.

The switch flushed out a latent guard gap: f32 sibling-subtraction noise
pushes degenerate cuts (every row one side) past the gain>0 gate; empty-
child covers then made SHAP contributions NaN. demote_empty_splits (and
the leafwise equivalent) now convert any split whose partition leaves a
child empty back to a leaf — row counts are ground truth at any precision.

+++ src/grower.cpp
@@ -227,8 +227,8 @@ void run_fill(FillPlan const &plan, Dataset const &ds, floats_view grad,
             {
                 size_t const         r  = rows[k];
                 uint8_t const *const rb = rm_ptr + (r * n_features);
-                double const         g  = grad[r];
-                double const         h  = hess[r];
+                float const          g  = grad[r];
+                float const          h  = hess[r];
                 for (size_t s = 0; s < n_sel; ++s)
                 {
                     HistCell &c = base_ptr[s][rb[sel_ptr[s]]];
